backend/requirements.txt

paho-mqtt==2.1.0
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
